refactor: make GridResourceAttributeKeys part of API
