+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/grid/GridResourceAttributeKeys.java
@@ -1,0 +1,13 @@
+package com.refinedmods.refinedstorage2.platform.api.grid;
+
+import com.refinedmods.refinedstorage2.api.grid.view.GridResourceAttributeKey;
+
+import org.apiguardian.api.API;
+
+@API(status = API.Status.STABLE, since = "2.0.0-milestone.2.6")
+public enum GridResourceAttributeKeys implements GridResourceAttributeKey {
+    MOD_ID,
+    MOD_NAME,
+    TAGS,
+    TOOLTIP
+}
